@@ -3,8 +3,8 @@
   "version": "0.0.0",
   "license": "MIT",
   "engines": {
-    "node": ">=18.10.0",
-    "npm": ">=8.19.0"
+    "node": ">=22.12.0",
+    "npm": ">=10.9.0"
   },
   "scripts": {
     "nx": "nx",
@@ -51,16 +51,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "17.3.7",
-    "@angular/cdk": "17.3.7",
-    "@angular/common": "17.3.7",
-    "@angular/compiler": "17.3.7",
-    "@angular/core": "17.3.7",
-    "@angular/forms": "17.3.7",
-    "@angular/material": "17.3.7",
-    "@angular/platform-browser": "17.3.7",
-    "@angular/platform-browser-dynamic": "17.3.7",
-    "@angular/router": "17.3.7",
+    "@angular/animations": "20.2.4",
+    "@angular/cdk": "20.2.2",
+    "@angular/common": "20.2.4",
+    "@angular/compiler": "20.2.4",
+    "@angular/core": "20.2.4",
+    "@angular/forms": "20.2.4",
+    "@angular/material": "20.2.2",
+    "@angular/platform-browser": "20.2.4",
+    "@angular/platform-browser-dynamic": "20.2.4",
+    "@angular/router": "20.2.4",
     "@types/trusted-types": "^2.0.2",
     "bootstrap": "5.2.3",
     "highlight.js": "^11.7.0",
@@ -125,15 +126,15 @@
     "karma-jasmine-html-reporter": "^2.0.0",
     "karma-mocha-reporter": "^2.2.5",
     "ng-mocks": "^14.5.1",
-    "ng-packagr": "17.3.0",
-    "nx": "18.3.4",
+    "ng-packagr": "20.2.0",
+    "nx": "21.5.1",
     "postcss": "^8.4.5",
     "postcss-import": "14.1.0",
     "postcss-preset-env": "7.5.0",
     "postcss-url": "10.1.3",
     "prettier": "3.2.5",
     "tailwindcss": "^3.2.4",
-    "typescript": "5.4.5",
+    "typescript": "5.9.2",
     "webpack-bundle-analyzer": "^4.8.0"
   },
   "config": {
